+++ lib/dealog_backoffice_web/live/components/banner_error.ex
@@ -1,0 +1,8 @@
+defmodule DealogBackofficeWeb.BannerError do
+  use DealogBackofficeWeb, :live_component
+
+  @impl true
+  def mount(socket) do
+    {:ok, socket}
+  end
+end