+++ lib/dealog_backoffice_web/router.ex
@@ -10,6 +10,15 @@ defmodule DealogBackofficeWeb.Router do
     plug :put_secure_browser_headers
   end
 
+  pipeline :preview do
+    plug :accepts, ["html"]
+    plug :fetch_session
+    plug :fetch_live_flash
+    plug :put_root_layout, {DealogBackofficeWeb.LayoutView, :preview}
+    plug :protect_from_forgery
+    plug :put_secure_browser_headers
+  end
+
   pipeline :api do
     plug :accepts, ["json"]
   end
@@ -26,6 +35,13 @@ defmodule DealogBackofficeWeb.Router do
     live("/my-account", MyAccountLive, :index)
     live("/settings", SettingsLive, :index)
     live("/readme", ReadmeLive, :index)
+    live("/design-system", DesignSystemLive, :index)
+  end
+
+  scope "/_preview", DealogBackofficeWeb do
+    pipe_through :preview
+
+    live("/", PreviewLive)
   end
 
   # Other scopes may use custom stacks.